// src/types/slices/profileSlice.ts
export interface ProfileState {
    profile: any;
    loading: boolean;
    error: string | null;
    updateProfileSuccess: any;
    updateProfileLoading: boolean;
    updateProfileError: string | null;
    updateProfilePasswordSuccess: any;
    updateProfilePasswordLoading: boolean;
    updateProfilePasswordError: string | null;
  }
  